cleaned up  some unused code, changed the level gen parameters slightly.
