+++ Assets/Scenes/Game.unity
@@ -794,7 +794,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   RoomData: {fileID: 11400000, guid: da5038c4ca23a4d4ea21eee5f9703979, type: 2}
-  LevelDimensions: {x: 5, y: 6}
+  LevelDimensions: {x: 6, y: 6}
   openDoorChance: 60
 --- !u!1001 &9167530697498633460
 PrefabInstance:
